Resolve "Make posts linkable"

See merge request cse2115/2020-2010/3-forum-discussion/op10-sem20!56

+++ contentserver/src/main/java/nl/tudelft/sem/group20/contentserver/repositories/ThreadRepository.java
@@ -4,7 +4,7 @@ import java.util.Optional;
 import nl.tudelft.sem.group20.contentserver.entities.BoardThread;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ThreadRepository extends JpaRepository<BoardThread, Integer> {
+public interface ThreadRepository extends JpaRepository<BoardThread, Long> {
 
     Optional<BoardThread> getById(long id);
 }
